[refactor] animator
- rename count to animationCount
- create delegate
- remove completion handler
