+++ NaverMapA/NaverMapA/Animator/AnimationType.swift
@@ -1,0 +1,13 @@
+//
+//  AnimationType.swift
+//  NaverMapA
+//
+//  Created by 채훈기 on 2020/12/14.
+//
+
+import Foundation
+
+enum AnimationType {
+    case move
+    case appear
+}
